In SitesService replaced find() function by query property and added on sitesChanged event.
The sitesChanged event is used in the minimap to sync footprints with filtered sites.
Fixes #43

In the searchbox the SitesService.query property is used directly, without need for on change watches.

Removed unused SitesService.getAllFeatures function.

Setting SitesService.query to a value does the following steps:
1. Update filtered and searched site list properties
2. Publishes the sitesChanged event
3. Triggers a digest to let angular detect changes in the filtered and searched site list properties

After sites are loaded the sitesChanged event is also published.
Fixes #44