Changed minimap extent fitting behaviour
In an attempt to make minimap controller more testable, I decided that
it's better to take the fitting of the map to a certain extent out of
the two functions that were using this and to put it into one (testable)
function. After this, it became trivial to also fit the extent of the
map to *both* the frustum *and* the (searched) sites. This improves the
minimap considerably; now one can see at the same time the frustum and
the sites, so that one can navigate towards them. Before, the map would
center only on either the sites *or* the frustum, making navigating
difficult if they weren't coincidentally already close to each other
(and thus on the map at the same time).